refactor option model
